added handle_put module
