+++ app/routing/convert_to_object.js
@@ -14,11 +14,5 @@ function convertValueArrayAndCategoriesToObject(valueArray, categories) {
     }
     return valueObject;
 }
-// function randomizeSelection(choicesObject) {
-//     let index = Math.floor(Math.random() * Object.keys(choicesObject).length);
-//     let selection = choicesObject[index];
-//     console.log(selection);
-//     return selection;
-// }
 
 export default convertValueArrayAndCategoriesToObject;
