+++ app/routing/restaurant_activity.js
@@ -58,19 +58,4 @@ hangoutRouter.route('/activity')
         }).catch(err => {console.log(err)});
     });
 
-// function convertValueArrayAndCategoriesToObject(valueArray, categories) {
-//     let valueObject = {};
-//     for (let i = 0; i < categories.length; i++) {
-//         valueObject[categories[i]] = valueArray[i];
-//     }
-//     return valueObject;
-// }
-// function randomizeSelection(choicesObject) {
-//     let index = Math.floor(Math.random() * Object.keys(choicesObject).length);
-//     let selection = choicesObject[index];
-//     console.log(selection);
-//     return selection;
-// }
-
-
 export default restaurantActivity;
